Moved protocols to CurlRuntime
